+++ noruba/features/lib/support/junit.rb
@@ -1,0 +1,7 @@
+module JUnitWorld
+  def replace_junit_time(time)
+    time.gsub(/\d+\.\d\d+/m, '0.05')
+  end
+end
+
+World(JUnitWorld)
